예외처리 추가 22/08/23 오후 4시 47분

+++ src/main/java/com/example/spring_jwt/repository/CommentRepository.java
@@ -3,6 +3,6 @@ package com.example.spring_jwt.repository;
 import com.example.spring_jwt.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
 }
